+++ src/view/task-edit/templates/task-color/task-color.ts
